New: Complete rewrite of FixedHeader.
Rather than showing cloned elements, it will now move the "real"
elements into floating containers and clone placeholders into the
original table to stop it jumping around.

The big benefit of this is that the original elements are the ones which
are on display to the end user all the time, so complex behaviour such
as putting inputs into the elements will still work. As a trivial
example the sort listeners attached by DataTables to the column header
automatically work since it is the original header that is shown.

There are also a number of performance benefits in doing it this way
since column width and height equalisation is not required.

I have removed the ability to fix left and right columns into place. The
overhead for that was too great and FixedColumns or a Responsive table
should be used instead.

This rewrite specifically targets DataTables 1.10+