Dev: Reworking how column alignment works
This is a frustrating change since it adds the `_matchWidths` function
back into FixedHeader, rather than using the `colgroup/col` tags from
DataTables 2.

The reason for that is if the table is actually drawn wider than what is
calculated by DataTables (which can happen due to "mmm" being wider than
"iiii" (non-mono-space!), then we'd end up with misalignment since
FixedHeader would show the colgroup/col sizes, while the main DataTable
would be slightly streached by the content.

Moveover, I've hadd to account for the padding-right on the table header
when scrolling and scrollbars are shown, otherwise we'd get misalignment
when we scroll to the right extreme.

Is there a better way? I don't know! Sum the colgroup/col and check
against the table width, if that is different then use this method,
otherwise use the clone of the colgroup? That's just more code and
another place for things to go wrong!