Dev: Third attempt at column alignment. colgroup / col with measured widths
The attempt to copy widths for cells just completely fails for complex
headers where a column doesn't have a unique header cell. Like not even
close, and interestingly fails in the same way in FF and Chrome: https://live.datatables.net/kofedida/1/edit

After a lot of head scratching I've gone back to the colgroup approach,
but to update the widths the `col`s define with the measured widths from
the DOM. This appears to be the most reliable method.